Started working on trainers page, created components folder to separate components from pages
