+++ client/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "axios": "^1.8.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
