+++ client/src/components/TrainerRequest.jsx
@@ -1,0 +1,10 @@
+import { useState } from 'react';
+
+export default function TrainerRequest() {
+    const [dialog, setDialog] = useState(true);
+    return(
+        <div className="trainer-request-container">
+
+        </div>
+    );
+}
