+++ client/src/pages/TrainersPage.jsx
@@ -1,12 +1,32 @@
+import { useEffect, useState } from 'react';
 import '../styles/TrainersPage.css';
-import NavBar from './NavBar';
+import NavBar from '../components/NavBar';
+import axios from 'axios';
+import { Link } from 'react-router-dom';
 
 export default function TrainersPage() {
-    return(
+    const [trainers, setTrainers] = useState([]);
+
+    useEffect(() => {
+        axios.get('http://localhost:3000/api/users/role/trainer')
+            .then((response) => {
+                setTrainers(response.data);
+            })
+            .catch((error) => {
+                console.log(error);
+            });
+    }, []);
+    return (
         <div className='trainers-page'>
-            <NavBar/>
+            <NavBar />
             <section>
-                <h1>Test</h1>
+                {
+                    trainers.map((trainer) => (
+                        <ul className='trainers-list' key={trainer.id}>
+                            <li>{trainer.name} <Link className='request-workout'>Request Workout</Link></li>
+                        </ul>
+                    ))
+                }
             </section>
         </div>
     );
