+++ client/src/styles/TrainersPage.css
@@ -1,0 +1,31 @@
+.trainers-list {
+    display: flex;
+    list-style: none;
+    gap: 10px;
+    justify-content: center;
+    margin-top: 100px;
+}
+
+.trainers-list li {
+    display: flex;
+    flex-direction: column;
+    gap: 30px;
+    padding: 10px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    align-items: center;
+    height: 100px;
+    width: 150px;
+}
+
+.trainers-list li .request-workout {
+    background-color: #4CAF50;
+    color: white;
+    padding: 10px 20px;
+    text-align: center;
+    text-decoration: none;
+    display: inline-block;
+    font-size: 16px;
+    border-radius: 5px;
+    cursor: pointer;
+}
